[ShuffleLog] Split TaskManager file view
