+++ ShuffleLog/src/main/java/com/swrobotics/shufflelog/tool/taskmanager/file/FileEditorTool.java
@@ -1,4 +1,4 @@
-package com.swrobotics.shufflelog.tool.taskmanager;
+package com.swrobotics.shufflelog.tool.taskmanager.file;
 
 import com.swrobotics.messenger.client.MessengerClient;
 import com.swrobotics.shufflelog.ShuffleLog;
