+++ ShuffleLog/src/main/java/com/swrobotics/shufflelog/tool/taskmanager/file/RemoteDirectory.java
@@ -1,4 +1,4 @@
-package com.swrobotics.shufflelog.tool.taskmanager;
+package com.swrobotics.shufflelog.tool.taskmanager.file;
 
 import java.util.Collection;
 import java.util.LinkedHashMap;
