+++ ShuffleLog/src/main/java/com/swrobotics/shufflelog/tool/taskmanager/file/RemoteFile.java
@@ -1,4 +1,4 @@
-package com.swrobotics.shufflelog.tool.taskmanager;
+package com.swrobotics.shufflelog.tool.taskmanager.file;
 
 public final class RemoteFile extends RemoteNode {
     public RemoteFile(String name) {
